Create grading_students.py
REF: https://www.hackerrank.com/challenges/grading/submissions/code/44591639

Initialize an empty array result which will hold the array
for the final grades. Iterate through grades. Then for each
grade, check if it's lower than 38. If it is, then it remains
unchanged and is passed immediately to the result array.
Else, pass the grade to the round_up function then append
to the result array.

To get the next multiple of 5 closest to the given grade,
the round_up function will be initialized to 40 then using
a while loop will be incremented by 5 until it is greater
than the given grade. Then a simple if check to see if the
grade should be replaced by the next_multiple.